requirement.txt
adjustText==0.6.1
aiodns==1.1.1
aiohttp==2.3.7
appnope==0.1.0
async-timeout==2.0.0
awscli==1.14.30
beautifulsoup4==4.6.0
bleach==2.1.2
Boruta==0.1.5
botocore==1.8.34
bs4==0.0.1
cchardet==2.1.1
ccxt==1.10.658
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.7
comtypes==1.1.4
cycler==0.10.0
dask==0.16.1
decorator==4.1.2
docutils==0.14
docxtpl==0.4.6
entrypoints==0.2.3
ghostscript==0.4.1
google-api-python-client==1.6.5
html5lib==1.0.1
httplib2==0.10.3
idna==2.6
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
lightgbm==2.0.12
lxml==4.1.1
MarkupSafe==1.0
matplotlib==2.0.0
mistune==0.8.3
multidict==3.3.2
nbconvert==5.3.1
nbformat==4.4.0
networkx==1.11
nltk==3.2.5
notebook==5.4.0
numpy==1.11.3
oauth2client==4.1.2
pandas==0.19.2
pandoc==1.0.2
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
ply==3.10
prompt-toolkit==1.0.15
psycopg2==2.7.1
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycares==2.3.0
Pygments==2.2.0
PyMySQL==0.8.0
pyparsing==2.2.0
python-crontab==2.2.8
python-dateutil==2.6.1
python-docx==0.8.6
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.4
qtconsole==4.3.1
reportlab==3.4.0
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.12
scikit-image==0.12.3
scikit-learn==0.18.1
scipy==0.18.1
seaborn==0.7.1
Send2Trash==1.4.2
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
textblob==0.15.1
toolz==0.9.0
tornado==4.5.3
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.3
xgboost==0.6a2
xlrd==1.1.0
yarl==0.18.0
